Rename `server-top` -> `top-server`.
...so it follows the usual convention, and also matches `top-client`.
