+++ local-modules/@bayou/main-server/package.json
@@ -9,6 +9,6 @@
   "dependencies": {
     "@bayou/config-common-default": "local",
     "@bayou/config-server-default": "local",
-    "@bayou/server-top": "local"
+    "@bayou/top-server": "local"
   }
 }
