+++ local-modules/@bayou/top-server/README.md
@@ -1,4 +1,4 @@
-@bayou/server-top
+@bayou/top-server
 =================
 
 This module contains most of the top level server code, including parsing of
