+++ local-modules/@bayou/top-server/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@bayou/server-top",
+  "name": "@bayou/top-server",
   "version": "1.0.0",
 
   "dependencies": {
